+++ OUDS/Core/Themes/Wireframe/Tests/Providers/ComponentTokens/WireframeThemeBadgeComponentTokensProviderTests.swift
@@ -1,0 +1,39 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+@testable import OUDSThemesWireframe
+import Testing
+
+// swiftlint:disable type_name
+
+/// Check if the component tokens provider under tests has the default semantic token providers with the expected types.
+struct WireframeThemeBadgeComponentTokensProviderTests {
+
+    private var provider: WireframeThemeBadgeComponentTokensProvider
+
+    init() {
+        provider = WireframeThemeBadgeComponentTokensProvider()
+    }
+
+    // MARK: - Tests
+
+    @Test func defaultSpaceSemanticTokensProvider() throws {
+        #expect(provider.spaces is WireframeThemeSpaceSemanticTokensProvider)
+    }
+
+    @Test func defaultDimensionsSemanticTokensProvider() throws {
+        #expect(provider.dimensions is WireframeThemeDimensionSemanticTokensProvider)
+    }
+}
+
+// swiftlint:enable type_name