+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Wireframe theme (Orange-OpenSource/ouds-ios#669)
 - Text input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Text area component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Select input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
@@ -23,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Color semantic tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
 - Border semantic tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
 - Text input component tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
 - Button component tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
@@ -46,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- List item component tokens
 - Input text component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Select component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Orange theme font family raw tokens